Реализовал сервис билетов и сделал редизайн

+++ routes/excursionsList.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const { body, validationResult } = require('express-validator')
-const { updateExcursion, deleteExcursion, deletePicture } = require('../controllers/excursions')
+const { updateExcursion, deleteExcursion, deletePicture, deleteTicket } = require('../controllers/excursions')
 const { Excursions } = require('../models/excursions')
 const { Companies } = require('../models/companies')
 const multer = require('multer')
@@ -47,7 +47,7 @@ router.get('/',
 
 // Установим алиас для каждой экскурсии
 router.param('slug', async function (req, res, next, slug) {
-    req.excursion = await Excursions.findOne({ slug: slug })
+    req.excursion = await Excursions.findOne({ slug: slug }).populate('tickets')
     next()
 })
 
@@ -62,11 +62,11 @@ router.route('/:slug')
             res.render('excursionsForm', {
                 action: '/excursions-list/' + req.excursion.slug,
                 title: 'Редактирование экскурсии: ' + req.excursion.title,
-                data: { body: req.excursion, companies: companies, selected: req.excursion.company, pictures: req.excursion.picturesURLs, tags: req.excursion.tags },
+                data: { body: req.excursion, companies: companies, selected: req.excursion.company, pictures: req.excursion.picturesURLs, tags: req.excursion.tags, tickets: req.excursion.tickets },
                 success: {
                     isSuccess: false,
                     msg: ''
-                }, 
+                },
                 errors: {},
                 user: username
             })
@@ -80,9 +80,9 @@ router.route('/:slug')
 router.route('/:slug')
     .post(upload.array('pictures'),
         body('title').notEmpty().withMessage('Название экскурсии обязательно к заполнению'),
-        body('price').notEmpty().withMessage('Стоимость обязательна к заполнению').isNumeric().withMessage('Стоимость должна быть числом'),
-        body('description').notEmpty().withMessage('Описание обязательно к заполнению').isLength({min: 5 ,max: 50}).withMessage('Количество символов должно быть от 5 до 50'),
+        body('description').notEmpty().withMessage('Описание обязательно к заполнению').isLength({ min: 5, max: 50 }).withMessage('Количество символов должно быть от 5 до 50'),
         body('isApproved').toBoolean(),
+        body('tickets').notEmpty().withMessage('Добавьте по крайней мере один билет'),
 
         function (req, res) {
             let username = req.user ? req.user.username : 'guest'
@@ -93,7 +93,7 @@ router.route('/:slug')
                     res.render('excursionsForm', {
                         action: '/excursions-list/' + req.excursion.slug,
                         title: 'Редактирование ' + req.excursion.title,
-                        data: { body: req.body, companies: companies, selected: req.excursion.company },
+                        data: { body: req.body, companies: companies, selected: req.excursion.company, tickets: req.excursion.tickets },
                         errors: errors.array(),
                         success: {
                             isSuccess: false,
@@ -136,25 +136,42 @@ router.route('/:slug/delete')
         })
 
 
-// Удаления изображения
 router.route('/:slug')
     .delete(async function (req, res) {
 
         if (req.isAuthenticated()) {
             let result = {}
 
-            try {
-                result = await deletePicture(req.body.index, req.excursion.slug)
-            }
-            catch (err) {
-                console.error(err)
-            }
+            // Удаления изображения
+            if (req.body.action === 'deleteImg') {
+                try {
+                    result = await deletePicture(req.body.index, req.excursion.slug)
+                }
+                catch (err) {
+                    console.error(err)
+                }
 
-            if (result.success) {
-                res.status(200).json({ success: true, error: {} })
+                if (result.success) {
+                    res.status(200).json({ success: true, error: {} })
+                }
+                else {
+                    res.status(200).json({ success: false, error: result.error })
+                }
             }
-            else {
-                res.status(200).json({ success: false, error: result.error })
+            if (req.body.action === 'deleteTicket') {
+                try {
+                    result = await deleteTicket(req.body.id)
+                }
+                catch (err) {
+                    console.error(err)
+                }
+
+                if (result.success) {
+                    res.status(200).json({ success: true, error: {} })
+                }
+                else {
+                    res.status(200).json({ success: false, error: result.error })
+                }
             }
         }
         else {
